server/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg